Update README.md
Briefly summarize The Gaming Room client and their software requirements. Who was the client? What type of software did they want you to design? 

CLIENT SIDE: 

The application must be delivered as a modern, responsive HTML interface running inside the web browser for desktop clients (Linux, Mac, Windows), as well as on mobile platforms. Each will be capable of communicating with the back-end web application running on the server. It is important for the client to notice all the considerations dealing with each platform such as costs, time, and expertise. The client side is the running script on the client's device, in this case it is the person playing the game. We are Producing a web-based game for multiple platforms called “Draw It or Lose It” that is currently only on Android. 

 

 

What did you do particularly well in developing this documentation? 

I feel like I did a great job organizing and describing enough information to help others understand and learn the differences in the platforms. 

 

What about the process of working through a design document did you find helpful when developing the code? 

Working through a design document helped me stay organized and on track of what we are trying to accomplish. It also broke down information that helped me understand a little more and see what each platform had that may be similar or different. 

 

If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it? 

I feel like I could have laid out my recommendations a little better and organized it in a way that would be easier for someone that didn’t know much about this project to absorb more information and maybe even get a little more detailed. 

How did you interpret the user’s needs and implement them into your software design? Why is it so important to consider the user’s needs when designing? 

I explained each platform with their: 

Characteristics: Key characteristics of each operating platform such as Performance, stability, security, and difficulty. 

Advantages: Advantages that each operating platform holds that can help identify what each platform has to offer. 

Weaknesses: The limitations of each platform. 

Server- based Deployment: note whether each platform offers a server-based deployment method and how they will handle the required scale for multiple players. 

Potential Licensing Costs: Initial cost, ongoing licensing fees, and any other cost information needed to know for each server operating system of each platform.  

Costs: The client will have to look at the Developing and testing for each platform, although there might be similarities between the platforms there may be additional resource and efforts that are required from the developing team that can impact the costs. Setting a budget for the developing team and requesting frameworks such as Cross- platform development can keep the costs a reduced amount. 

Time: The time will be based on Complexity of each platform and requirments that the development team need to complete the tasks. So, the timeline needs room in case of problems between multiple platforms. 

Expertise: Each operating platform requires different skills and complexity that makes them different that the developers will need experience on to help create and ensure that is works correctly. 

Knowing this information helps identify what we will be working with through each platform and where things might be similar or different as we cross through each platform. It might also give a better understanding of where to start and what to expect.